Initial refactoring to use Vagrant plugins
This demo is using many custom written moving parts. While they help
achieve a very automated build, they conflate the ideas of automating
Puppet installations and the core intentions of this demonstration.

There have been very effective and easily re-usable Vagrant plugins
written for handling a large bulk of what custom scripts perform in
this demonstration.

This commit attempts to begin a basic refactor to utilize the Vagrant
plugins `vagrant-pe_build` and `vagrant-hosts` to simplify the non-core
details of this demonstration.